src/diagram2graph/FigAnalysis/ShapeExtraction/requirements.txt
# pip install -r requirements.txt
numpy==1.16.2
opencv-contrib-python==4.0.0.21
pillow==5.4.1
imutils==0.5.2
scipy==1.2.1
scikit-image==0.14.2
pytesseract==0.2.6
pyspellchecker==0.4.0